pyproject.toml

[project]
name = "llmanonymizationuv"
version = "0.1.0"
description = "LLM Anonymization Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diffprivlib>=0.6.5",
    "langchain[groq]>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-text-splitters>=0.3.6",
    "langgraph>=0.3.5",
    "ollama>=0.4.7",
    "openai>=1.65.4",
    "pip>=25.0.1",
    "presidio-analyzer>=2.2.357",
    "presidio-anonymizer>=2.2.357",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "spacy>=3.8.4",
    "streamlit>=1.43.0",
    "langchain-ollama>=0.2.3",
    "langchain-chroma>=0.2.2",
    "beautifulsoup4>=4.13.3",
    "langchain-openai>=0.3.7",
    "streamlit-sortables>=0.3.1",
    "ragas>=0.2.14",
    "sacrebleu>=2.5.1",
    "spacy-transformers>=1.3.8",
    "tqdm>=4.67.1",
    "torch>=2.6.0",
    "seqeval>=1.2.2",
    "scikit-learn>=1.6.1",
    "datasets>=3.3.2",
    "accelerate>=1.5.2",
    "pandas>=2.2.3",
    "ipykernel>=6.29.5",
    "pypdf>=5.4.0",
    "opacus>=1.5.3",
    "transformers>=4.49.0",
    "sentence-transformers>=4.0.1",
    "openpyxl>=3.1.5",
    "ragatouille>=0.0.9",
    "tenseal>=0.3.16",
    "cryptography>=44.0.2",
]
